scraper/requirements.txt
# Amazon Product Advertising API integration
python-amazon-paapi>=5.0.0

# Configuration and settings management
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Async HTTP client for web scraping
httpx>=0.27.0

# HTML parsing for web scraping fallback
beautifulsoup4>=4.12.0

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Type checking and linting
mypy>=1.0.0
ruff>=0.1.0

# Core dependencies
pydantic>=2.0.0
typing-extensions>=4.0.0

# Logging and utilities
loguru>=0.7.0

# Date and time handling
python-dateutil>=2.8.0

# JSON handling
orjson>=3.9.0